Graphical changes and fine tunings

+++ src/maps/Block.java
@@ -7,7 +7,7 @@ import org.newdawn.slick.geom.Polygon;
 public class Block {
 	public Polygon blockPoly;
 	public String blockName;
-	public Block(int x, int y, int points[], String type){
+	public Block(float x, float y, int points[], String type){
 		blockPoly = new Polygon(new float[]{x + points[0], y + points[1],
 											x + points[2], y + points[3],
 											x + points[4], y + points[5],
